Streamlit-Subapase-Db-Project/requirements.txt
streamlit
supabase
python-dotenv
psycopg2